refactor: @param: 어노테이션 제거 및 unsignedByte/unsignedShort 비교 코드 수정
- 테스트 클래스의 @Autowired 프로퍼티 선언에 @param: 어노테이션 제거
- unsignedByte 및 unsignedShort 비교 코드에서 명시적으로 `.toUByte()`, `.toUShort()` 적용
- 코드 가독성 개선 및 타입 명확성 확보

+++ 05-exposed-dml/02-types/src/test/kotlin/exposed/examples/types/Ex07_UnsignedColumnType.kt
@@ -160,7 +160,8 @@ class Ex07_UnsignedColumnType: JdbcExposedTestBase() {
                 val number2 = (Byte.MAX_VALUE + 1).toUByte()
                 assertFailAndRollback("Out-of-range (OoR) error") {
                     UByteTable.insert { it[unsignedByte] = number2 }
-                    UByteTable.selectAll().where { UByteTable.unsignedByte less 0u }.count().toInt() shouldBeEqualTo 0
+                    UByteTable.selectAll().where { UByteTable.unsignedByte less 0u.toUByte() }.count()
+                        .toInt() shouldBeEqualTo 0
                 }
 
                 // modify column to now have SMALLINT type
@@ -239,7 +240,7 @@ class Ex07_UnsignedColumnType: JdbcExposedTestBase() {
                 val number2 = (Short.MAX_VALUE + 1).toUShort()
                 assertFailAndRollback("Out-of-range (OoR) error") {
                     UShortTable.insert { it[unsignedShort] = number2 }
-                    UShortTable.selectAll().where { UShortTable.unsignedShort less 0u }.count()
+                    UShortTable.selectAll().where { UShortTable.unsignedShort less 0u.toUShort() }.count()
                         .toInt() shouldBeEqualTo 0
                 }
